refactor: extract EditWordDialog confirm mapping, split edit-dialog tests
- Pull the item.copy(...) construction on Save into a local
  applyEditedFields() helper, mirroring how AddWordUiState.toCardOptions()
  isolates the same mapping for the Add flow.
- Move all EditWordDialog tests (including the new bidirectional ones)
  into a dedicated EditWordDialogTest.kt, dropping WordListScreenTest.kt
  back under detekt's LargeClass threshold without a suppression.
